feat: 预防攻击

+++ src/controller/blog.js
@@ -1,11 +1,14 @@
-const { exec } = require('../db/mysql');
+const { exec, escape } = require('../db/mysql');
+const xss = require('xss')
 const bloglist = (author, keyword) => {
     let sql = `select * from blogs where 1=1 `
     if (author) {
+        author = escape(author)
         sql += `and author = ${author} `;
     }
     if (keyword) {
-        sql += `and keyword = ${keyword} `;
+        keyword = escape(keyword)
+        sql += `and title like '%${keyword}%' `;
     }
     sql += `order by createtime desc`
 
@@ -13,16 +16,19 @@ const bloglist = (author, keyword) => {
     return exec(sql)
 }
 const blogDetail = (id) => {
-    let sql = `select * from blogs where id = ${id}`;
-    return exec(sql);
+    let sql = `select * from blogs where id=${id}`;
+    return exec(sql).then(rows => {
+        return rows[0];
+    });
 }
 const blogAdd = (blogData = {}) => {
-    const title = blogData.title;
-    const content = blogData.content;
+    const title = xss(blogData.title);
+    const content = xss(blogData.content);
     const createtime = Date.now();
-    const author = 'zhangsan';
+    const author = escape(blogData.author);
+    // const author = 'zhangsan';
     // 作者应该是从登陆的地方拿到的
-    let sql = `insert into blogs(title,content,createtime,author) values ('${title}','${content}','${createtime}','${author}')`
+    let sql = `insert into blogs(title,content,createtime,author) values (${title},${content},'${createtime}',${author})`;
     return exec(sql).then(res => {
         return {
             id: res.insertId
@@ -30,9 +36,9 @@ const blogAdd = (blogData = {}) => {
     });
 }
 const blogUpdate = (id, blogData = {}) => {
-    const title = blogData.title;
-    const content = blogData.content;
-    let sql = `update blogs set title='${title}', content='${content}' where id=${id};`
+    const title = xss(blogData.title);
+    const content = xss(blogData.content);
+    let sql = `update blogs set title=${title},content=${content} where id=${id}`;
     return exec(sql).then(res => {
         if (res.affectedRows > 0) {
             return true
@@ -40,8 +46,9 @@ const blogUpdate = (id, blogData = {}) => {
         return false
     })
 }
-const blogDelete = (id) => {
-    let sql = `delete from blogs where id=${id}`
+const blogDelete = (id, author) => {
+    author = escape(author);
+    let sql = `delete from blogs where id=${id} and author=${author}`
     return exec(sql).then(res => {
         if (res.affectedRows > 0) {
             return true
